feat(frontend): refactor tables to Server Components with streaming and Suspense
