+++ frontend/app/(core)/operations/customers/_components/table.tsx
@@ -1,27 +1,14 @@
-'use client';
-
 import { CustomerDto } from '@sgcv2/shared';
 import { CustomerDropMenu } from './customerDropMenu';
 import { statusMap } from '../_const/const';
 import { Badge } from '@/components/ui/badge';
 import { DataTable, Column } from '@/components/table/data-table';
-import { deleteCustomerAction } from './actions';
-import { toast } from 'sonner';
 
 interface CustomersTableProps {
   data: CustomerDto[];
 }
 
 export function CustomersTable({ data }: CustomersTableProps) {
-  const handleDelete = async (id: string) => {
-    const result = await deleteCustomerAction(id);
-    if (result.success) {
-      toast.success('Cliente eliminado exitosamente');
-    } else {
-      toast.error(result.message || 'Error al eliminar el cliente');
-    }
-  };
-
   const columns: Column<CustomerDto>[] = [
     {
       header: 'Código',
@@ -54,11 +41,7 @@ export function CustomersTable({ data }: CustomersTableProps) {
       columns={columns}
       emptyMessage="No se encontraron resultados."
       rowActions={customer => (
-        <CustomerDropMenu
-          id={customer.id}
-          customerName={customer.legalName}
-          onDelete={handleDelete}
-        />
+        <CustomerDropMenu id={customer.id} customerName={customer.legalName} />
       )}
     />
   );
