+++ frontend/app/(core)/operations/customers/page.tsx
@@ -1,8 +1,8 @@
-import { serverCustomersService } from '@/lib/api/server-customers.service';
-import { CustomersTable } from './_components/table';
 import { CustomersFilters } from './_components/filters';
-import { CustomerDto, CustomerState } from '@sgcv2/shared';
-import { DataPagination } from '@/components/data-pagination';
+import { CustomerState } from '@sgcv2/shared';
+import { Suspense } from 'react';
+import { CustomersTableContent } from './_components/table-content';
+import { TableSkeleton } from '@/components/table/table-skeleton';
 
 interface CustomersPageProps {
   searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
@@ -18,45 +18,6 @@ export default async function CustomersPage({ searchParams }: CustomersPageProps
     state: typeof params.status === 'string' ? (params.status as CustomerState) : undefined,
   };
 
-  let customers: CustomerDto[] = [];
-  let totalPages = 1;
-
-  try {
-    const response = await serverCustomersService.getAll(page, perPage, filters);
-    if (response.success) {
-      customers = response.data || [];
-      // Note: Assuming the service or a future update provides actual total pages
-      // For now, if we have a way to calculate it or if metadata exists.
-      // Based on users page pattern:
-      totalPages = response.metadata?.pagination?.totalPages || 1;
-    }
-  } catch (error) {
-    console.error('Error fetching customers:', error);
-  }
-
-  const createPageUrl = (newPage: number) => {
-    const urlParams = new URLSearchParams();
-
-    if (filters.search) {
-      urlParams.set('search', filters.search);
-    }
-
-    if (filters.state) {
-      urlParams.set('status', filters.state);
-    }
-
-    if (newPage > 1) {
-      urlParams.set('page', newPage.toString());
-    }
-
-    if (perPage !== 5) {
-      urlParams.set('perPage', perPage.toString());
-    }
-
-    const queryString = urlParams.toString();
-    return queryString ? `?${queryString}` : '/operations/customers';
-  };
-
   return (
     <div className="p-6 space-y-4">
       <div className="flex flex-col gap-2">
@@ -65,9 +26,9 @@ export default async function CustomersPage({ searchParams }: CustomersPageProps
 
       <CustomersFilters search={filters.search} status={filters.state} />
 
-      <CustomersTable data={customers} />
-
-      <DataPagination currentPage={page} totalPages={totalPages} createPageUrl={createPageUrl} />
+      <Suspense key={JSON.stringify(params)} fallback={<TableSkeleton columnCount={5} />}>
+        <CustomersTableContent page={page} perPage={perPage} filters={filters} />
+      </Suspense>
     </div>
   );
 }
