+++ frontend/app/(core)/users/page.tsx
@@ -1,8 +1,8 @@
-import { serverUsersService } from '@/lib/api/server-users.service';
-import { UsersTable } from '../../../feature/users/components/admin/users-table';
-import { UsersFilters } from '../../../feature/users/components/admin/filters';
-import { UserDto, UserFilterDto, UserStatus } from '@sgcv2/shared';
-import { DataPagination } from '@/components/data-pagination';
+import { UsersFilters } from '@feature/users/components/admin/filters';
+import { UserFilterDto, UserStatus } from '@sgcv2/shared';
+import { Suspense } from 'react';
+import { UsersTableContent } from './_components/table-content';
+import { TableSkeleton } from '@/components/table/table-skeleton';
 import { Button } from '@/components/ui/button';
 import Link from 'next/link';
 
@@ -24,44 +24,6 @@ export default async function UsersPage({ searchParams }: UsersPageProps) {
     },
   };
 
-  let users: UserDto[] = [];
-  let totalPages = 1;
-  const currentPage = Math.floor(offset / limit) + 1;
-
-  try {
-    const response = await serverUsersService.getAll(filter);
-    if (response.success) {
-      users = response.data || [];
-      totalPages = response.metadata?.pagination?.totalPages || 1;
-    }
-  } catch (error) {
-    console.error('Error fetching users:', error);
-  }
-
-  const createPageUrl = (page: number) => {
-    const newOffset = (page - 1) * limit;
-    const searchParams = new URLSearchParams();
-
-    if (params.search && params.search.toString() !== '') {
-      searchParams.set('search', params.search.toString());
-    }
-
-    if (params.status && params.status.toString() !== '') {
-      searchParams.set('status', params.status.toString());
-    }
-
-    if (newOffset > 0) {
-      searchParams.set('offset', newOffset.toString());
-    }
-
-    if (limit !== 20) {
-      searchParams.set('limit', limit.toString());
-    }
-
-    const queryString = searchParams.toString();
-    return queryString ? `?${queryString}` : '/users';
-  };
-
   return (
     <div className="p-6 space-y-6">
       <header className="flex justify-between items-center">
@@ -77,13 +39,9 @@ export default async function UsersPage({ searchParams }: UsersPageProps) {
       <main className="flex w-full flex-col gap-4">
         <UsersFilters search={filter.search} status={filter.status} />
 
-        <UsersTable data={users} />
-
-        <DataPagination
-          currentPage={currentPage}
-          totalPages={totalPages}
-          createPageUrl={createPageUrl}
-        />
+        <Suspense key={JSON.stringify(params)} fallback={<TableSkeleton columnCount={4} />}>
+          <UsersTableContent limit={limit} offset={offset} filter={filter} />
+        </Suspense>
       </main>
     </div>
   );
