+++ frontend/components/table/table-skeleton.tsx
@@ -1,0 +1,43 @@
+import { Skeleton } from '@/components/ui/skeleton';
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table';
+
+interface TableSkeletonProps {
+  columnCount: number;
+  rowCount?: number;
+}
+
+export function TableSkeleton({ columnCount, rowCount = 5 }: TableSkeletonProps) {
+  return (
+    <div className="rounded-md border">
+      <Table>
+        <TableHeader>
+          <TableRow>
+            {Array.from({ length: columnCount }).map((_, i) => (
+              <TableHead key={i}>
+                <Skeleton className="h-4 w-[100px]" />
+              </TableHead>
+            ))}
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {Array.from({ length: rowCount }).map((_, i) => (
+            <TableRow key={i}>
+              {Array.from({ length: columnCount }).map((_, j) => (
+                <TableCell key={j}>
+                  <Skeleton className="h-4 w-full" />
+                </TableCell>
+              ))}
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </div>
+  );
+}
