+++ frontend/feature/users/components/admin/users-table.tsx
@@ -1,15 +1,10 @@
-'use client';
-
 import { UserDto } from '@sgcv2/shared';
 import { UserDropMenu } from './userDropMenu';
 import { Badge } from '@/components/ui';
 import { DataTable, type Column } from '@/components/table/data-table';
-import { blockUserAction } from '@feature/users/actions';
-import { toast } from 'sonner';
 
 interface UsersTableProps {
   data: UserDto[];
-  isLoading?: boolean;
 }
 
 const statusMap = {
@@ -18,16 +13,7 @@ const statusMap = {
   BLOCKED: { label: 'Bloqueado', variant: 'destructive' as const },
 };
 
-export function UsersTable({ data = [], isLoading }: UsersTableProps) {
-  const handleDelete = async (id: number) => {
-    const result = await blockUserAction(id);
-    if (result.success) {
-      toast.success('Usuario bloqueado correctamente');
-    } else {
-      toast.error(result.error?.message || 'Error al bloquear el usuario');
-    }
-  };
-
+export function UsersTable({ data = [] }: UsersTableProps) {
   const columns: Column<UserDto>[] = [
     {
       header: 'ID',
@@ -56,11 +42,8 @@ export function UsersTable({ data = [], isLoading }: UsersTableProps) {
     <DataTable
       data={data}
       columns={columns}
-      isLoading={isLoading}
       emptyMessage="No se encontraron usuarios."
-      rowActions={user => (
-        <UserDropMenu id={user.id} username={user.username} onDelete={handleDelete} />
-      )}
+      rowActions={user => <UserDropMenu id={user.id} username={user.username} />}
     />
   );
 }
